requirements.txt
fitz==0.0.1.dev2
Flask==2.3.2
msal==1.22.0
Pillow==9.5.0
Pillow==9.5.0
PyMuPDF==1.22.2
PyPDF2==3.0.1
python_pptx==0.6.21
Requests==2.31.0
